+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#2099** : Fix stepping source pane for segmented (cooked) data.
+
 * Add null protection to `login_count` and `login_failures` in `users` to `account` table migration.